requirements.txt
argon2-cffi==20.1.0
asgiref==3.4.0
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.3.0
boto3==1.17.109
botocore==1.20.109
bs4==0.0.1
cffi==1.14.4
colorama==0.4.4
comtypes==1.1.8
decorator==4.4.2
defusedxml==0.6.0
Django==3.2.4
django-crispy-forms==1.12.0
django-storages==1.11.1
entrypoints==0.3
gunicorn==20.1.0
ipykernel==5.4.3
ipython==7.20.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.3
jmespath==0.10.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.11
jupyter-console==6.2.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
MarkupSafe==1.1.1
mistune==0.8.4
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
notebook==6.2.0
packaging==20.9
pandocfilters==1.4.3
parso==0.8.1
pickleshare==0.7.5
Pillow==8.3.0
prometheus-client==0.9.0
prompt-toolkit==3.0.14
pycparser==2.20
Pygments==2.7.4
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
pyzmq==22.0.2
qtconsole==5.0.2
QtPy==1.9.0
s3transfer==0.4.2
Send2Trash==1.5.0
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
terminado==0.9.2
testpath==0.4.4
tornado==6.1
traitlets==5.0.5
urllib3==1.26.6
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
